refactor: split atom schemas
